--- src/main/java/com/delivery/finalproject/util/MyCommon.java
@@ -1,10 +1,0 @@
-package com.delivery.finalproject.util;
-
-public class MyCommon {
-
-	public static class Photo {
-
-		public static final int BLOCK_LIST = 8; // 한화면에 보여질 게시물수
-		public static final int BLOCK_PAGE = 5; // 한화면에 보여질 페이지수
-	}
-}
