--- src/main/java/com/delivery/finalproject/vo/DeliveryHistoryVo.java
@@ -1,13 +1,0 @@
-package com.delivery.finalproject.vo;
-
-import org.apache.ibatis.type.Alias;
-
-import lombok.Data;
-
-@Data
-@Alias("DeliveryHistory")
-public class DeliveryHistoryVo {
-    private int delivery_history_id;
-    private String delivery_history_status;
-    private int deliveries_id;
-}
