--- src/main/java/com/delivery/finalproject/vo/RiderVo.java
@@ -1,16 +1,0 @@
-package com.delivery.finalproject.vo;
-
-import org.apache.ibatis.type.Alias;
-
-import lombok.Data;
-
-@Data
-@Alias("rider")
-public class RiderVo {
-    private int raiders_id;
-    private String rider_name;
-    private String riders_phone;
-    private String riders_email;
-    private String riders_pwd;
-    private String riders_region;
-}
