+++ src/main/java/first/final_project/FinalProjectApplication.java
@@ -1,15 +1,15 @@
-package com.delivery.finalproject;
+package first.final_project;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@MapperScan(basePackages = "com.delivery.finalproject.mapper")
+@MapperScan(basePackages = "first.final_project.dao")
 public class FinalProjectApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(FinalProjectApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(FinalProjectApplication.class, args);
+    }
 
 }
