+++ src/main/java/first/final_project/vo/DeliveriesVo.java
@@ -1,0 +1,14 @@
+package first.final_project.vo;
+
+import org.apache.ibatis.type.Alias;
+
+import lombok.Data;
+
+@Data
+@Alias("deliveries")
+public class DeliveriesVo {
+    private int deliveries_id;
+    private int raiders_id;
+    private String deliveries_method;
+    private int orders_id;
+}
