+++ src/main/resources/application.properties
@@ -16,13 +16,9 @@ spring.datasource.driver-class-name: com.mysql.cj.jdbc.Driver
 #mybatis.configuration.map-underscore-to-camel-case=true
 
 #Model Location
-mybatis.type-aliases-package=com.delivery.finalproject.vo
+mybatis.type-aliases-package=first.final_project.vo
 #Xml Location
 mybatis.mapper-locations=classpath:mappers/*.xml
-
-# JSP 사용을 위한 설정 
-spring.mvc.view.prefix=/WEB-INF/views/
-spring.mvc.view.suffix=.jsp
 
 # LiveReload 브라우저 장프로그램 설치시 적용
 spring.devtools.livereload.enabled=true
