+++ src/main/resources/mappers/RiderMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.delivery.finalproject.mapper.RiderMapper">
+<mapper namespace="first.final_project.dao.RidersMapper">
 
     <!-- 전체조회 -->
     <select id="selectList" resultType="rider">
