+++ src/main/webapp/index.html
@@ -1,32 +1,14 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-pageEncoding="UTF-8"%>
 <!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <title>Insert title here</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
 
     <script>
-      // 메뉴이동 href
-      //   location.href = "menu/list.do";
-      // 가게이동 href
-      location.href = "shop/list.do";
+      //location.href="shop/list.do"
+      location.href = "riders/main.do";
     </script>
   </head>
-  <body>
-    ㅎ
-  </body>
-  <!DOCTYPE html>
-  <html lang="en">
-    <head>
-      <meta charset="UTF-8" />
-      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <title>Document</title>
-
-      <script>
-        location.href = "riders/main.do";
-      </script>
-    </head>
-    <body></body>
-  </html>
+  <body></body>
 </html>
